build: move build logic to bob package
Because both local and Jenkins builds use the same engine, move the code
into the bob package itself. No functional change intended.
